projects/trend-tracker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trend-tracker"
version = "0.1.0"
description = "News and tech trend tracking MCP server with snapshot comparison"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "supabase>=2.0",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/trend_tracker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
